requirements/dev.txt
pyspark
psycopg2-binary
python-dotenv
pytest
numpy<2
pandas<2
pyarrow
